+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Recipe App 1.1
+- added recipe deletion
+- added recipe portion count
+- added recipe instructions
+- slightly changed UI
+- updated current recipes
+
 # Recipe App 1.0
 - added recipe file saving and loading
 - added recipe creation